requirements.txt
hydra-core
omegaconf
h5py
numpy==1.26.0
numcodecs
opencv-python-headless==4.11.0.86
pandas
scipy
zarr
tqdm
einops
wandb
pynput
click
threadpoolctl
matplotlib
dill
pydantic
rtree
shapely
open3d
diffusers
pytorch-lightning
lightning-bolts
tyro
sentry-sdk==1.43.0
typing_extensions==4.12.2